refactor: 이미지 엔티티 관련 기능 수정
