+++ src/main/java/project/volunteer/domain/image/application/ImageServiceImpl.java
@@ -6,13 +6,11 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import project.volunteer.domain.image.dao.ImageRepository;
 import project.volunteer.domain.image.domain.Image;
-import project.volunteer.domain.image.domain.ImageType;
 import project.volunteer.domain.logboard.dao.LogboardRepository;
 import project.volunteer.global.common.component.RealWorkCode;
 import project.volunteer.domain.image.application.dto.ImageParam;
 import project.volunteer.domain.recruitment.dao.RecruitmentRepository;
-import project.volunteer.domain.storage.application.StorageService;
-import project.volunteer.domain.storage.domain.Storage;
+import project.volunteer.domain.image.domain.Storage;
 import project.volunteer.domain.user.dao.UserRepository;
 import project.volunteer.global.error.exception.BusinessException;
 import project.volunteer.global.error.exception.ErrorCode;
@@ -25,28 +23,20 @@ public class ImageServiceImpl implements ImageService{
 
     private final StorageService storageService;
     private final ImageRepository imageRepository;
-    private final UserRepository userRepository;
-    private final RecruitmentRepository recruitmentRepository;
-    private final LogboardRepository logboardRepository;
 
     @Transactional
     @Override
+    //TODO: 퍼사드 패턴 적용하기
     public Long addImage(ImageParam saveImageDto) {
-
-        //참조 엔티티 유무 검사필요(user, recruitment, log...)
-        validateNo(saveImageDto.getNo(), saveImageDto.getCode());
-
+        //이미지 타입 저장
         Image createImage = Image.builder()
                 .realWorkCode(saveImageDto.getCode())
                 .no(saveImageDto.getNo())
-                .staticImageName(saveImageDto.getStaticImageCode())
                 .build();
 
-        //업로드 이미지일 경우
-        if(saveImageDto.getImageType()== ImageType.UPLOAD){
-            Storage uploadImageStorage = storageService.addStorage(saveImageDto.getUploadImage(), saveImageDto.getCode());
-            createImage.setStorage(uploadImageStorage);
-        }
+        //이미지 정보 저장
+        Storage uploadImageStorage = storageService.addStorage(saveImageDto.getUploadImage(), saveImageDto.getCode());
+        createImage.setStorage(uploadImageStorage);
 
         return imageRepository.save(createImage).getImageNo();
     }
@@ -69,20 +59,4 @@ public class ImageServiceImpl implements ImageService{
         imageRepository.findImagesByCodeAndNo(code, no).stream()
                 .forEach(img -> img.setDeleted());
     }
-
-    private void validateNo(Long no, RealWorkCode code) {
-
-        //더 클린하게 작성할 수 없을까?
-        if(code==RealWorkCode.USER) {
-            userRepository.findById(no).orElseThrow(() ->
-                    new BusinessException(ErrorCode.NOT_EXIST_USER, String.format("User No = [%d]", no)));
-        }else if(code==RealWorkCode.RECRUITMENT){
-            recruitmentRepository.findById(no).orElseThrow(() ->
-                    new BusinessException(ErrorCode.NOT_EXIST_RECRUITMENT, String.format("Recruitment No = [%d]", no)));
-        }else if(code==RealWorkCode.LOG) {
-        	 logboardRepository.findById(no).orElseThrow(() ->
-                    new BusinessException(ErrorCode.NOT_EXIST_LOGBOARD, String.format("Logboard No = [%d]", no)));
-        }
-    }
-
 }
